pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-backend"
version = "1.0.0"
description = "Glimmora Aegis Aerospace — Training Platform Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "gunicorn>=21.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyjwt[crypto]>=2.8",
    "bcrypt>=4.1",
    "redis>=5.0",
    "celery>=5.3",
    "google-generativeai>=0.5",
    "openai>=1.20",
    "httpx>=0.27",
    "structlog>=24.1",
    "python-multipart>=0.0.9",
    "meilisearch-python-async>=1.8",
    "cryptography>=42.0",
    "python-jose[cryptography]>=3.3",
    "email-validator>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
